Properly extract class for submission state actions
h/t @damassi for the nudge
