+++ app/controllers/admin/submissions_controller.rb
@@ -1,71 +1,5 @@
 # frozen_string_literal: true
 
-class StateActions
-  def self.for(submission)
-    new(submission).run
-  end
-
-  attr_reader :submission
-
-  def initialize(submission)
-    @submission = submission
-  end
-
-  def run
-    actions = []
-    actions << approve_action if submission.submitted?
-    actions << publish_action if submission.submitted? || submission.approved?
-    actions << reject_action if submission.submitted?
-    actions << close_action if submission.submitted?
-    actions
-  end
-
-  private
-
-  def default_classes
-    ['btn btn-secondary btn-small btn-full-width']
-  end
-
-  def approve_action
-    {
-      class: default_classes << 'btn-approve',
-      confirm:
-        'An email will be sent to the consignor, letting them know that their submission will be sent to our partner network and this work will appear in the digests and CMS. This action cannot be undone.',
-      state: 'approved',
-      text: 'Approve (convection only)'
-    }
-  end
-
-  def publish_action
-    {
-      class: default_classes << 'btn-approve',
-      confirm:
-        'An email will be sent to the consignor, letting them know that their submission will be sent to our partner network and this work will appear in the digests and CMS. This action cannot be undone.',
-      state: 'published',
-      text: 'Publish (CMS + digest)'
-    }
-  end
-
-  def reject_action
-    {
-      class: default_classes << 'btn-delete',
-      confirm:
-        'An email will be sent to the consignor, letting them know that we are not accepting their submission. This action cannot be undone.',
-      state: 'rejected',
-      text: 'Reject'
-    }
-  end
-
-  def close_action
-    {
-      class: default_classes << 'btn-delete',
-      confirm: 'No email will be sent.',
-      state: 'closed',
-      text: 'Close'
-    }
-  end
-end
-
 module Admin
   class SubmissionsController < ApplicationController
     include GraphqlHelper
@@ -155,7 +89,7 @@ module Admin
       @partner_submissions_count =
         notified_partner_submissions.group_by_day.count
       @offers = @submission.offers
-      @actions = StateActions.for(@submission)
+      @actions = SubmissionStateActions.for(@submission)
     end
 
     def edit; end
